+++ examples/basic/main.go
@@ -38,13 +38,9 @@ func (c App) Validate() error {
 
 func main() {
 	app := &App{}
-	config, err := config.New(app,
-		config.WithLoader(config.EnvProvider[*App]("APP_", "__")),
-		config.WithLoader(config.FileProvider[*App]("./config/app.json")),
-	)
-	if err != nil {
-		panic(err)
-	}
+	config := config.New(app).
+		WithProvider(config.EnvProvider[*App]("APP_", "__")).
+		WithProvider(config.FileProvider[*App]("./config/app.json"))
 
 	ctx := context.Background()